webserver.socketio settings
