+++ services/web/server/src/simcore_service_webserver/socketio/_schema.py
@@ -1,0 +1,16 @@
+""" socketio subsystem's configuration
+
+    - config-file schema
+    - settings
+"""
+
+import trafaret as T
+
+CONFIG_SECTION_NAME = "socketio"
+
+
+schema = T.Dict(
+    {
+        T.Key("enabled", default=True, optional=True): T.Or(T.Bool(), T.ToInt()),
+    }
+)
